requirements.txt
#tkinter>=8.6
psycopg2-binary>=2.9.9
pandas>=2.0.0
geopandas>=0.14.0
shapely>=2.0.0
matplotlib>=3.7.0
descartes>=1.1.0
